+++ src/components/MindTable.vue
@@ -374,8 +374,6 @@ export default {
         path: '/articleMde',
         query: {
           selected: this.data[index].id,
-          source: 'article',
-          essay: this.data.note,
           pageCurrent: this.page.current,
         },
       });
